backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
anthropic==0.39.0
httpx==0.27.2
python-multipart==0.0.12
langdetect==1.0.9
